+++ seqkit/cmd/subseq.go
@@ -57,7 +57,7 @@ Examples:
 		lineWidth := config.LineWidth
 		outFile := config.OutFile
 		quiet := config.Quiet
-		seq.AlphabetGuessSeqLenghtThreshold = config.AlphabetGuessSeqLength
+		seq.AlphabetGuessSeqLengthThreshold = config.AlphabetGuessSeqLength
 		seq.ValidateSeq = false
 		gtf.Threads = config.Threads
 		fai.MapWholeFile = false